performance improvements
loss of precision from f64 to f32 is acceptable, large memory savings
from i64 to u32

+++ src/lib.rs
@@ -127,7 +127,7 @@ impl<T> Injector<T> {
 #[derive(PartialEq, Debug, Clone, Copy)]
 pub struct Match {
     pub score: u32,
-    pub frecency_score: f64,
+    pub frecency_score: f32,
     pub idx: u32,
 }
